reorganize files in components
