+++ src/components/AdminExam/ExameList.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Exame } from '../types/Exame';
+import { Exame } from '../../types/Exame.ts';
 import { Button, IconButton } from '@mui/material';
 import { InsertPhoto, ArrowForwardIos, KeyboardArrowDown  } from '@mui/icons-material';
