+++ src/components/Booking/BookingConfirmation.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import {Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle} from "@/components/ui/card.tsx";
 import {Button} from "@/components/ui/button.tsx";
-import {DadosPaciente} from "@/components/RegisterPatient.tsx";
-import {DadosBooking, Exams} from "@/components/Booking.tsx";
+import {DadosPaciente} from "@/components/AdminPatient/RegisterPatient.tsx";
+import {DadosBooking, Exams} from "@/components/Booking/Booking.tsx";
 import {CalendarDays, Clock, FileText, MapPin, User} from "lucide-react";
 
 export interface BookingConfirmationProps {
