+++ src/components/Booking/CheckCPF.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react'
-import { Button } from "@/components/ui/button"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
+import { Button } from "@/components/ui/button.tsx"
+import { Input } from "@/components/ui/input.tsx"
+import { Label } from "@/components/ui/label.tsx"
 import {
     Card,
     CardContent,
@@ -9,10 +9,10 @@ import {
     CardFooter,
     CardHeader,
     CardTitle,
-} from "@/components/ui/card"
+} from "@/components/ui/card.tsx"
 import {AlertCircle} from "lucide-react"
-import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert"
-import {DadosPaciente} from "@/components/RegisterPatient.tsx";
+import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert.tsx"
+import {DadosPaciente} from "@/components/AdminPatient/RegisterPatient.tsx";
 import {useAuth} from "@/hooks/auth.tsx";
 import {getPatientByCpfAndTenant} from "@/services/patientService.tsx";
 import {validarCPF} from "@/lib/utils.ts";
