+++ src/components/Booking/StepIndicator.tsx
